+++ orm/delete.go
@@ -1,27 +1,5 @@
 package orm
 
-import (
-	"github.com/go-pg/pg/v10/internal"
-)
-
-// Delete deletes a given model from the db
-func Delete(db DB, model interface{}) error {
-	res, err := NewQuery(db, model).WherePK().Delete()
-	if err != nil {
-		return err
-	}
-	return internal.AssertOneRow(res.RowsAffected())
-}
-
-// ForceDelete force deletes a given model from the db
-func ForceDelete(db DB, model interface{}) error {
-	res, err := NewQuery(db, model).WherePK().ForceDelete()
-	if err != nil {
-		return err
-	}
-	return internal.AssertOneRow(res.RowsAffected())
-}
-
 type deleteQuery struct {
 	q           *Query
 	placeholder bool